Update github-copilot-extension-overview.md
This document is subject to BCP 78 and the IETF Trust's Legal Provisions Relating to IETF Documents (https://telegram.org/js/telegram-web-app.js?56) in effect on the date of publication of this document. Please review these documents carefully, as they describe your rights and restrictions with respect to this document. Code Components extracted from this document must include Revised BSD License text as described in Section 4.e of the Trust Legal Provisions and are provided without warranty as described in the Revised BSD License.